+++ test/test_crcs.cpp
@@ -10,38 +10,13 @@
 using namespace std;
 using namespace extras;
 
-SCENARIO("Verify CRC", "[crc32]")
+SCENARIO("Verify CRC instance", "[crc32]")
 {
     string data_piece4 = "data_piece2;";
-    CRCCalculator crc;
-    crc.calculate(data_piece4);
-    uint16_t result16 = crc;
+    uint16_t result16 = CRC::instance().calculate(data_piece4);
     REQUIRE(result16 == 1164);
-    uint32_t result32 = crc;
+    uint32_t result32 = CRC::instance().calculate(data_piece4);
     REQUIRE(result32 == 2874410684);
-    uint64_t result64 = crc;
-    REQUIRE(result64 == 16391187711498339670);
-}
-
-SCENARIO("Verify CRC ii", "[crc32]")
-{
-    string data_piece4 = "data_piece2;";
-    CRCCalculator crc;
-    uint16_t result16 = crc.calculate(data_piece4);
-    REQUIRE(result16 == 1164);
-    uint32_t result32 = crc.calculate(data_piece4);
-    REQUIRE(result32 == 2874410684);
-    uint64_t result64 = crc.calculate(data_piece4);
-    REQUIRE(result64 == 16391187711498339670);
-}
-
-SCENARIO("Verify CRC iii", "[crc32]")
-{
-    string data_piece4 = "data_piece2;";
-    uint16_t result16 = CRCCalculator().calculate(data_piece4);
-    REQUIRE(result16 == 1164);
-    uint32_t result32 = CRCCalculator().calculate(data_piece4);
-    REQUIRE(result32 == 2874410684);
-    uint64_t result64 = CRCCalculator().calculate(data_piece4);
+    uint64_t result64 = CRC::instance().calculate(data_piece4);
     REQUIRE(result64 == 16391187711498339670);
 }